requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
python-dotenv>=1.0
groq>=0.4
psycopg2-binary>=2.9
django-cors-headers>=4.0
gunicorn>=21.0